# ProductService/src/main/resources/api.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8092/product/service","description":"Generated server url"}],"paths":{"/cart":{"post":{"tags":["cart-resource"],"operationId":"addItemToCart","parameters":[{"name":"SESSION_TOKEN","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShoppingCartItemRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShoppingCartItemDto"}}}}}}}},"/session/public":{"get":{"tags":["session-resource"],"operationId":"getPublicSessionToken","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/product/{productId}":{"get":{"tags":["product-resource"],"operationId":"getProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}}}}},"/product/prototype":{"get":{"tags":["product-resource"],"operationId":"saveProduct","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/category/{categoryId}":{"get":{"tags":["product-resource"],"operationId":"getCategory","parameters":[{"name":"categoryId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}}}},"/cart/{sessionKey}":{"get":{"tags":["cart-resource"],"operationId":"getItemsFromCart","parameters":[{"name":"sessionKey","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShoppingCartItemDto"}}}}}}}}},"components":{"schemas":{"ShoppingCartItemRequest":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"}}},"ShoppingCartItemDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"sessionKey":{"type":"string"},"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"modifiedAt":{"type":"string","format":"date-time"}}},"CategoryDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"categoryName":{"type":"string"},"categoryPath":{"type":"string"},"parent":{"$ref":"#/components/schemas/CategoryDto"},"children":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDto"}},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}}}},"ProductAttributeDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"value":{"type":"string"}}},"ProductDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productName":{"type":"string"},"retailPrice":{"type":"number"},"discountedPrice":{"type":"number"},"imageLink":{"type":"string"},"category":{"$ref":"#/components/schemas/CategoryDto"},"relatedProductDto":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/RelatedProductDto"}},"productAttributes":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/ProductAttributeDto"}}}},"RelatedProductDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productName":{"type":"string"}}},"ProductAttributeResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"value":{"type":"string"}}},"ProductCategoryResponse":{"type":"object"},"ProductResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productName":{"type":"string"},"retailPrice":{"type":"number"},"discountedPrice":{"type":"number"},"imageLink":{"type":"string"},"category":{"$ref":"#/components/schemas/ProductCategoryResponse"},"relatedProductDto":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/RelatedProductResponse"}},"productAttributes":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/ProductAttributeResponse"}}}},"RelatedProductResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productName":{"type":"string"}}}}}}